requirements.txt
azure-storage-blob
pandas
pyarrow
flask
python-dotenv
gunicorn